# api/openapi/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package openapi

import (
	"github.com/go-faster/errors"
)

// Ref: #/components/schemas/Error
type Error struct {
	Code ErrorCode `json:"code"`
}

// GetCode returns the value of Code.
func (s *Error) GetCode() ErrorCode {
	return s.Code
}

// SetCode sets the value of Code.
func (s *Error) SetCode(val ErrorCode) {
	s.Code = val
}

// Ref: #/components/schemas/ErrorCode
type ErrorCode string

const (
	ErrorCodeAccountsUnavailable ErrorCode = "accounts_unavailable"
	ErrorCodeInvalidRequest      ErrorCode = "invalid_request"
)

// AllValues returns all ErrorCode values.
func (ErrorCode) AllValues() []ErrorCode {
	return []ErrorCode{
		ErrorCodeAccountsUnavailable,
		ErrorCodeInvalidRequest,
	}
}

// MarshalText implements encoding.TextMarshaler.
func (s ErrorCode) MarshalText() ([]byte, error) {
	switch s {
	case ErrorCodeAccountsUnavailable:
		return []byte(s), nil
	case ErrorCodeInvalidRequest:
		return []byte(s), nil
	default:
		return nil, errors.Errorf("invalid value: %q", s)
	}
}

// UnmarshalText implements encoding.TextUnmarshaler.
func (s *ErrorCode) UnmarshalText(data []byte) error {
	switch ErrorCode(data) {
	case ErrorCodeAccountsUnavailable:
		*s = ErrorCodeAccountsUnavailable
		return nil
	case ErrorCodeInvalidRequest:
		*s = ErrorCodeInvalidRequest
		return nil
	default:
		return errors.Errorf("invalid value: %q", data)
	}
}

type GetLoginBadRequest Error

func (*GetLoginBadRequest) getLoginRes() {}

type GetLoginServiceUnavailable Error

func (*GetLoginServiceUnavailable) getLoginRes() {}

// Ref: #/components/schemas/Login
type Login struct {
	CanLogin     bool   `json:"can_login"`
	ClientName   string `json:"client_name"`
	RedirectHost string `json:"redirect_host"`
	Users        []User `json:"users"`
}

// GetCanLogin returns the value of CanLogin.
func (s *Login) GetCanLogin() bool {
	return s.CanLogin
}

// GetClientName returns the value of ClientName.
func (s *Login) GetClientName() string {
	return s.ClientName
}

// GetRedirectHost returns the value of RedirectHost.
func (s *Login) GetRedirectHost() string {
	return s.RedirectHost
}

// GetUsers returns the value of Users.
func (s *Login) GetUsers() []User {
	return s.Users
}

// SetCanLogin sets the value of CanLogin.
func (s *Login) SetCanLogin(val bool) {
	s.CanLogin = val
}

// SetClientName sets the value of ClientName.
func (s *Login) SetClientName(val string) {
	s.ClientName = val
}

// SetRedirectHost sets the value of RedirectHost.
func (s *Login) SetRedirectHost(val string) {
	s.RedirectHost = val
}

// SetUsers sets the value of Users.
func (s *Login) SetUsers(val []User) {
	s.Users = val
}

func (*Login) getLoginRes() {}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/User
type User struct {
	Email string `json:"email"`
	// Optional account identifier displayed instead of email.
	Identifier OptString `json:"identifier"`
	Name       string    `json:"name"`
	Subject    string    `json:"subject"`
}

// GetEmail returns the value of Email.
func (s *User) GetEmail() string {
	return s.Email
}

// GetIdentifier returns the value of Identifier.
func (s *User) GetIdentifier() OptString {
	return s.Identifier
}

// GetName returns the value of Name.
func (s *User) GetName() string {
	return s.Name
}

// GetSubject returns the value of Subject.
func (s *User) GetSubject() string {
	return s.Subject
}

// SetEmail sets the value of Email.
func (s *User) SetEmail(val string) {
	s.Email = val
}

// SetIdentifier sets the value of Identifier.
func (s *User) SetIdentifier(val OptString) {
	s.Identifier = val
}

// SetName sets the value of Name.
func (s *User) SetName(val string) {
	s.Name = val
}

// SetSubject sets the value of Subject.
func (s *User) SetSubject(val string) {
	s.Subject = val
}